function/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.17.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-http-client==3.3.7
sendgrid==6.10.0
starkbank-ecdsa==2.2.0
